LayoutTests/fast/hidpi/image-srcset-fraction-1.5x.html

<html>
<head>
<script>
    window.targetScaleFactor = 1.5;
</script>
<script src="resources/srcset-helper.js"></script>
<script src="../../resources/js-test-pre.js"></script>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }

    addEventListener("load", function() {
        if (internals)
            shouldBeTrue('document.getElementById("testimg").clientWidth==200');
    }, false);
</script>
</head>
<body>
<div>This test passes if the srcset resource is loaded and displayed as the image</div>
<img src="resources/blue-100-px-square.png" srcset="resources/green-400-px-square.png 2x" id="testimg">
</body>
</html>
